Move classes to top level
