+++ transactionoutbox-core/src/main/java/com/gruelbox/transactionoutbox/SQLAction.java
@@ -1,0 +1,9 @@
+package com.gruelbox.transactionoutbox;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+
+@FunctionalInterface
+interface SQLAction {
+  void doAction(Connection connection) throws SQLException;
+}
